新增我的模块部分页面

+++ aiiopenapi/src/main/java/com/aiitec/openapi/constant/AIIConstant.java
@@ -7,7 +7,7 @@ public class AIIConstant {
     public static String FILTER_ACTION_LOGIN_ON_OTHER = "com.aiitec.openapi.login_on_other";
     public static String EXTRA_STATUS = "status";
 
-    public static long USER_ID ;
+    public static int USER_ID ;
     public static String AII_APP_ID = "123";
     public static String AII_APP_KEY = "123456";
 }
